refactor byline into react component
Refactors the byline into its own React Component. In general, I think this helps reduce interpretability of the code. No changes in UI. Minimal changes to implementation, related to making this a functional component which are simpler than react classes (and react is generally moving away from classes).
